pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wildguard"
version = "0.1.0"
description = "Scalable Oversight Harness for Dark-Pattern Monitoring in Real LLM Chat Logs"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "WildGuard Team"}
]
keywords = ["llm", "monitoring", "dark-patterns", "nlp", "oversight"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "datasets>=2.14.0",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "streamlit>=1.28.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
]

[project.scripts]
wildguard-ingest = "src.ingest_wildchat:main"
wildguard-judge = "src.judge_label:main"
wildguard-train = "src.train_classifier:main"
wildguard-infer = "src.infer_wildchat:main"
wildguard-report = "src.analytics:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
